Moved exporter methods to own class
